pyproject.toml
[project]
name = "gen2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.14"
dependencies = [
    "anthropic>=0.37.1",
    "bitsandbytes>=0.45.0",
    "cyclopts>=3.16.1",
    "datasets>=3.2.0",
    "debugpy>=1.8.8",
    "docker",
    "fastapi>=0.115.12",
    "genson>=1.3.0",
    "google-generativeai>=0.8.5",
    "gradio>=5.25.2",
    "jinja2>=3.1.4",
    "jsonlines",
    "loguru",
    "ollama>=0.4.4",
    "openai",
    "pandas>=2.2.3",
    "pip>=24.3.1",
    "plotext",
    "polars>=1.19.0",
    "prometheus-client>=0.21.0",
    "psutil",
    "pydantic",
    "pyright>=1.1.400",
    "python-configuration>=0.12.1",
    "python-dotenv",
    "python-on-whales>=0.76.1",
    "pyyaml>=6.0.2",
    "result>=0.17.0",
    "schedule>=1.2.2",
    "scikit-learn>=1.6.1",
    "setuptools>=75.8.0",
    "transformers>=4.47.1",
    "trl>=0.13.0",
    "unidecode>=1.3.8",
    "unsloth>=2025.1.6",
    "uvicorn>=0.34.2",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.scripts]
main_script = "scripts:main"

[tool.setuptools]
py-modules = ['scripts', 'src']

[tool.pyright]
include = [
    "scripts/*.py",
    "src/**/*.py"
]
exclude = [
    "**/__pycache__",
    ".venv/**",
    "scripts/unused/*.py"
]
reportMissingImports = "warning"  # Change to "none" if you want to ignore those
verboseOutput = true
